+++ src/main/java/com/zhiyin/plugins/oneClickNavigation/codeInsight/JavaToMocLineMarkerProvider.java
@@ -44,7 +44,7 @@ public class JavaToMocLineMarkerProvider extends RelatedItemLineMarkerProvider {
             NavigationGutterIconBuilder<PsiElement> builder;
             builder = NavigationGutterIconBuilder
                     .create(MyIcons.pandaIconSVG16_2)
-                    .setAlignment(GutterIconRenderer.Alignment.RIGHT)
+                    .setAlignment(GutterIconRenderer.Alignment.LEFT)
                     .setTargets(targets)
                     .setTooltipText(Constants.NAVIGATE_TO_MOC);
             RelatedItemLineMarkerInfo<PsiElement> relatedItemLineMarkerInfo = builder.createLineMarkerInfo(element);